On/Off Monitor now quits when it is shut down
I was going to upgrade socket to serversocket, which can be shut down, but I noticed that Thread objects have a daemon attribute. When this is set to True, the Thread is stopped when no more non-daemon threads remain alive. socketserver was difficult to get going anyway (it's just TCP, like socket, which is why the browser didn't like it when the server just responded with text)